Move supports_relative template to right place.
